Add files via upload
Abstract – This paper explores the use of the simulation 
model in Simply 4.4 to simulate a motorway. The 
motorway simulated follows the traffic laws and rules of 
speed limit for different type of vehicles in Ireland.

Conclusions
To summarize the answer to the original research 
questions, it can be said that, in according to the 
parameters given and the simulations done, the presence 
of 1% of Tesla in the motorway contributes to decrease 
the number of crashes, the introduction of a speed limit 
in the merging area across the lanes does not improve the 
quality of the traffic, but it actually worsens it, and the 
best scenario for the traffic is given with a percentage of 
Tesla of 1% and a IAT of 0.52.